fix(core): avoid Turbopack circular dependency
