+++ packages/core/src/dom/media-actions.ts
@@ -1,8 +1,8 @@
-import { isUndefined } from '@videojs/utils/predicate';
-
-import { DEFAULT_SEEK_STEP, DEFAULT_VOLUME_STEP } from '../core/ui/constants';
+import { getMediaInputActionValue } from './media-action-value';
 import type { AnyPlayerStore } from './player';
 import { selectPlaybackRate, selectTime, selectVolume } from './store/selectors';
+
+export { getMediaInputActionValue } from './media-action-value';
 
 export type MediaInputActionName = 'seekStep' | 'volumeStep' | 'speedUp' | 'speedDown';
 
@@ -13,28 +13,6 @@ export interface MediaInputActionContext {
 }
 
 export type MediaInputActionResolver = (context: MediaInputActionContext) => void;
-
-export function getMediaInputActionValue(
-  action: string,
-  key: string | undefined,
-  value?: number | undefined
-): number | undefined {
-  if (!isUndefined(value)) return value;
-
-  const normalizedKey = key?.toLowerCase();
-
-  if (action === 'seekStep') {
-    return normalizedKey === 'arrowleft' || normalizedKey === 'j' ? -DEFAULT_SEEK_STEP : DEFAULT_SEEK_STEP;
-  }
-
-  if (action === 'volumeStep') {
-    const step = DEFAULT_VOLUME_STEP / 100;
-
-    return normalizedKey === 'arrowdown' ? -step : step;
-  }
-
-  return undefined;
-}
 
 export const MEDIA_INPUT_ACTION_OVERRIDES: Record<MediaInputActionName, MediaInputActionResolver> = {
   seekStep({ store, value, key }) {
